# specs/001-create-a-rest/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Test Results Management API
  description: |
    REST API for CRUD operations on end-to-end test execution results.
    Primary integration via GitHub Actions CI/CD pipelines using bearer token authentication.

    ## Authentication
    All endpoints require JWT bearer token authentication. Tokens are obtained through
    GitHub SSO for human users or generated for automated systems (CI/CD pipelines).

    ## Integration Patterns
    - **GitHub Actions**: Use bearer tokens stored in repository/organization secrets
    - **CI/CD Systems**: Automated test result submission with artifact upload
    - **Development**: Local testing with personal access tokens
  version: 0.4.0
  contact:
    name: Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.testresults.dev
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestSuite:
      type: object
      required: [name, framework_id, environment_id, total_tests, passed_tests, failed_tests, skipped_tests, duration_ms, started_at, completed_at]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 200
          example: "Login Flow Tests"
        framework_id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        total_tests:
          type: integer
          minimum: 0
          example: 25
        passed_tests:
          type: integer
          minimum: 0
          example: 20
        failed_tests:
          type: integer
          minimum: 0
          example: 3
        skipped_tests:
          type: integer
          minimum: 0
          example: 2
        duration_ms:
          type: integer
          minimum: 0
          example: 45000
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    PaginatedTestSuites:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestSuite'
        total:
          type: integer
          example: 50
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        pages:
          type: integer
          example: 3
...
paths:
...
  /suites:
    get:
      summary: List test suites
      tags: [Suites]
      parameters:
        - name: framework_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by framework
        - name: environment_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by environment
        - name: name
          in: query
          schema:
            type: string
          description: Filter by suite name
        - name: started_after
          in: query
          schema:
            type: string
            format: date-time
          description: Filter suites started after this date
        - name: started_before
          in: query
          schema:
            type: string
            format: date-time
          description: Filter suites started before this date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of test suites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTestSuites'

    post:
      summary: Create test suite
      tags: [Suites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestSuite'
      responses:
        '201':
          description: Suite created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSuite'
        '400':
          description: Invalid input
  /suites/{suite_id}:
    get:
      summary: Get test suite
      tags: [Suites]
      parameters:
        - name: suite_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Suite details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSuite'
        '404':
          description: Suite not found

    put:
      summary: Update test suite
      tags: [Suites]
      parameters:
        - name: suite_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestSuite'
      responses:
        '200':
          description: Suite updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSuite'

    delete:
      summary: Delete test suite
      tags: [Suites]
      parameters:
        - name: suite_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Suite deleted successfully
        '404':
          description: Suite not found

  # Test Results
...